Replaced the <head> of experience.html
I tried to replicate the resolution that I was made aware of by replacing the head from the experience.html file with lines 1-8 from the index.html file. However I still received an error when I checked my work using the W3C CSS Validation Service. I'm not sure what a "parse error" is and why I keep getting them. I'm not sure if this is significant or not but I validated my CSS using the "direct input tab"